fazendo rodar

+++ include/utility/heap.h
@@ -68,7 +68,7 @@ public:
         int * addr = reinterpret_cast<int *>(e->object() + e->size());
 
         if(typed)
-            *addr++ = reinterpret_cast<int>(this);
+            *addr++ = reinterpret_cast<unsigned long>(this);
         *addr++ = bytes;
 
         db<Heaps>(TRC) << ") => " << reinterpret_cast<void *>(addr) << endl;
